Refactor: PD-246 이메일 필드 분리
